+++ src/ZeppelinForms/Forms/Controls/Charts/PieSlice.cs
@@ -1,0 +1,10 @@
+﻿using ZeppelinForms.Drawing.Primitives;
+
+namespace ZeppelinForms.Forms.Controls.Charts;
+
+public sealed class PieSlice
+{
+    public string? Label { get; set; }
+    public float Value { get; set; }
+    public Color? Color { get; set; }
+}